some more gutting, reduce warnings
whoever thought it was fine to write nearly 200 NullPointerExceptions needs to be taken to an insane asylum. I'm not fixing all that shit. goddamn 2k+ warnings brought down to just under 600.

+++ Essentials/src/com/earth2me/essentials/storage/EnchantmentLevel.java
@@ -55,11 +55,12 @@ public class EnchantmentLevel implements Entry<Enchantment, Integer> {
     @Override
     public boolean equals(final Object obj) {
         if (obj instanceof Entry) {
+            //noinspection rawtypes
             final Entry entry = (Entry) obj;
             if (entry.getKey() instanceof Enchantment && entry.getValue() instanceof Integer) {
                 final Enchantment enchant = (Enchantment) entry.getKey();
                 final Integer lvl = (Integer) entry.getValue();
-                return this.enchantment.equals(enchant) && this.level == lvl.intValue();
+                return this.enchantment.equals(enchant) && this.level == lvl;
             }
         }
         return false;
